move weekly roundup implementation from script into package to facilitate tests
